refactor: extract Template Studio interactions
