+++ apps/web/src/routes/_protected.knowledge/-components/template-studio.tsx
@@ -1,24 +1,11 @@
-import type { MouseEvent as ReactMouseEvent, ReactNode } from "react";
 import { lazy, Suspense, useCallback, useMemo, useRef, useState } from "react";
 
 import { useQuery, useQueryClient } from "@tanstack/react-query";
-import type { LucideIcon } from "lucide-react";
-import {
-  BracesIcon,
-  CheckIcon,
-  ChevronDownIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  RepeatIcon,
-  SplitIcon,
-  TextQuoteIcon,
-  WandSparklesIcon,
-} from "lucide-react";
+import { BracesIcon, RepeatIcon, SplitIcon } from "lucide-react";
 import type { NodeType, Node as PMNode, ResolvedPos } from "prosemirror-model";
 import type { EditorState, Transaction } from "prosemirror-state";
 import { TextSelection } from "prosemirror-state";
 import type { EditorView } from "prosemirror-view";
-import { useDebounce, useDebouncedCallback } from "use-debounce";
 import { useTranslations } from "use-intl";
 
 import type { TemplateRecipeDefinition } from "@stll/api/types";
@@ -27,50 +14,26 @@ import type {
   DirectiveRange,
   DocxEditorRef,
   TemplatePreviewValue,
-  TemplateSlashMenuKeyAction,
-  TemplateSlashMenuState,
 } from "@stll/folio-react";
 import {
   buildPositionalText,
-  clearTemplateSlashMenu,
-  consumeTemplateSlashQuery,
-  getFolioCaretViewportRect,
-  getFolioSelectionViewportRect,
   getTemplateDirectives,
-  getTemplateSlashMenu,
-  resetTemplateSlashQuery,
   setTemplatePreviewValues,
 } from "@stll/folio-react";
 import "@stll/folio-react/editor.css";
-import {
-  isClauseSlotName,
-  isFieldPath,
-  isSafeFieldPath,
-} from "@stll/template-conditions";
+import { isClauseSlotName, isFieldPath } from "@stll/template-conditions";
 import { Button } from "@stll/ui/components/button";
-import { DirectionalIcon } from "@stll/ui/components/directional-icon";
-import {
-  MenuPreviewLayout,
-  PreviewPane,
-} from "@stll/ui/components/preview-pane";
-import { Separator } from "@stll/ui/components/separator";
 import { stellaToast } from "@stll/ui/components/toast";
-import { containedHandler } from "@stll/ui/hooks/use-contained-handler";
-import { cn } from "@stll/ui/lib/utils";
 
 import { useInspectorStore } from "@/components/inspector/inspector-store";
 import { useExternalSyncEffect, useMountEffect } from "@/hooks/use-effect";
-import type { TranslationKey } from "@/i18n/types";
-import { getAnalytics } from "@/lib/analytics/provider";
 import { api } from "@/lib/api";
 import { optionalArray } from "@/lib/arrays";
-import { BoundedMap } from "@/lib/bounded-set";
 import { DOCX_MIME } from "@/lib/consts";
 import { userErrorMessage } from "@/lib/errors/user-safe";
 import { toSafeId } from "@/lib/safe-id";
-import { inputTypeValueKind, VALUE_TYPE_META } from "@/lib/value-types";
+import { inputTypeValueKind } from "@/lib/value-types";
 import { TemplateStudioChat } from "@/routes/_protected.knowledge/-components/template-studio-chat";
-import { reusableConditions } from "@/routes/_protected.knowledge/-components/template-studio-condition-source";
 import "@/routes/_protected.knowledge/-components/template-studio-inspector";
 import {
   protectedRouteApi,
@@ -80,15 +43,21 @@ import {
 } from "@/routes/_protected.knowledge/-components/template-studio-constants";
 import {
   buildManifest,
-  isInputType,
   nextFreePath,
   parseFields,
   prepareRecipeInsert,
-  sanitizeFieldPath,
   slugify,
 } from "@/routes/_protected.knowledge/-components/template-studio-model";
 import { buildOutline } from "@/routes/_protected.knowledge/-components/template-studio-outline";
 import { useFitToWidth } from "@/routes/_protected.knowledge/-components/template-studio-preview";
+import {
+  TemplateStudioSelectionGesture,
+  useTemplateStudioSelectionGesture,
+} from "@/routes/_protected.knowledge/-components/template-studio-selection-gesture";
+import {
+  TemplateStudioSlashMenu,
+  useTemplateStudioSlashMenu,
+} from "@/routes/_protected.knowledge/-components/template-studio-slash-menu";
 import {
   defaultStudioField,
   useTemplateStudioStore,
@@ -96,9 +65,7 @@ import {
   type StudioField,
 } from "@/routes/_protected.knowledge/-components/template-studio-store";
 import { filledByForFieldMeta } from "@/routes/_protected.knowledge/-components/template-studio-suggestions";
-import type { TemplateEditableField } from "@/routes/_protected.knowledge/-components/template-wizard";
 import {
-  clausesOptions as clauseLibraryOptions,
   knowledgeKeys,
   templateDocxBufferOptions,
 } from "@/routes/_protected.knowledge/-queries";
@@ -149,26 +116,6 @@ const flashDirectiveAt = (view: EditorView, pos: number) => {
   requestAnimationFrame(run);
 };
 
-// ── Selection gesture popover ────────────────────────────
-
-/** Selecting prose settles for this long before the popover shows; drag and
- *  shift+arrow selections re-arm it instead of flashing mid-gesture. */
-const GESTURE_SHOW_DELAY_MS = 150;
-/** Pause on a stable selection before asking the model to enrich the
- *  Make-field row. The instant buttons never wait for this. */
-const GESTURE_ENRICH_DELAY_MS = 500;
-const GESTURE_POPOVER_OFFSET_PX = 8;
-/** Half the popover's widest layout (menu column + 256px preview pane, capped
- *  at the 30rem/480px max-width), for clamping its centered position inside the
- *  host so it never spills out. */
-const GESTURE_POPOVER_HALF_WIDTH_PX = 240;
-/** Rough rendered height of the menu + preview, for the above/below flip. */
-const GESTURE_POPOVER_EST_HEIGHT_PX = 280;
-/** At most this many existing field paths ride along in the enrichment
- *  instruction (the suggest endpoint bounds `instructions` at 2000 chars). */
-const GESTURE_ENRICH_MAX_PATHS = 30;
-/** Joined-paths character budget inside the enrichment instruction. */
-const GESTURE_ENRICH_PATHS_MAX_CHARS = 1000;
 /** Source-phrase cap for the bilingual-mirror instruction (the suggest
  *  endpoint bounds `instructions` at 2000 chars). */
 const MIRROR_SOURCE_MAX_CHARS = 400;
@@ -179,244 +126,13 @@ const MIRROR_OFFER_TOAST_MS = 10_000;
  *  convention (bare `{{lawyer}}` inside its own loop never substitutes). */
 const LOOP_ITEM_KEY = "value";
 
-/** The live text selection in the document, as reported by Folio. */
-type GestureSelection = { from: number; to: number; text: string };
-
-/** A settled selection the gesture popover is anchored to; `left`/`top` are
- *  relative to the page's document wrapper (the popover's offset parent). */
-type SelectionGesture = GestureSelection & {
-  left: number;
-  top: number;
-  placement: "above" | "below";
-};
-
-// ── Slash-command menu ───────────────────────────────────
-
-/** Gap (px) between the caret and the menu's top-left corner. Kept tiny so the
- *  menu hugs the cursor. */
-const SLASH_MENU_OFFSET_PX = 2;
-/** Rough rendered height of the menu, for the above/below flip and the
- *  off-screen clamp. */
-const SLASH_MENU_EST_HEIGHT_PX = 280;
-/** Full rendered menu width, for the off-screen right-edge clamp. The list
- *  column is `w-56` (224px); on `sm`+ screens `MenuPreviewLayout` renders a
- *  `PreviewPane` (`w-64`, 256px) beside it across a 9px gap+border, so the menu
- *  is ~489px wide. Clamp against the whole width so the preview pane never
- *  spills off-screen. */
-const SLASH_MENU_LIST_WIDTH_PX = 224;
-const SLASH_MENU_PREVIEW_WIDTH_PX = 256 + 9;
-const SLASH_MENU_WIDTH_PX =
-  SLASH_MENU_LIST_WIDTH_PX + SLASH_MENU_PREVIEW_WIDTH_PX;
-/** Tailwind `sm`: below it `MenuPreviewLayout` hides the preview pane, so only
- *  the list column renders and the right-edge clamp must use the list width. */
-const SLASH_MENU_PREVIEW_BREAKPOINT_PX = 640;
-/** Cap on existing fields offered for reuse so a large template's field list
- *  never turns the menu into an unbounded scroll. */
-const SLASH_MENU_MAX_FIELDS = 50;
-/** Cap on clause-library rows fetched for the clause submenu. */
-const SLASH_MENU_CLAUSE_LIMIT = 50;
-
-/** The open slash menu, positioned at the caret. `left`/`top` are relative to
- *  the overlay host (the menu's offset parent), like the gesture popover. */
-type SlashMenu = {
-  from: number;
-  query: string;
-  left: number;
-  top: number;
-  placement: "above" | "below";
-};
-
-/** Which level of the menu is showing. The root lists the creators and the two
- *  submenu openers; `fields` and `clauses` are the reuse submenus. */
-type SlashView = "root" | "fields" | "clauses";
-
-/** A top-level row: two direct creators and two submenu openers. */
-type SlashRootItem =
-  | { kind: "create-field"; path: string }
-  | { kind: "create-condition" }
-  | { kind: "field"; path: string; label: string }
-  | { kind: "open-fields" }
-  | { kind: "open-clauses" };
-
-/** A clause-library row in the slash menu's clause submenu. */
-type SlashClause = {
-  id: string;
-  title: string;
-  currentVersion: number;
-  description: string | null;
-};
-
-/** The rows for the active view, tagged so render and the key handler narrow
- *  on `view` without unsafe casts. */
-type SlashRows =
-  | { view: "root"; items: SlashRootItem[] }
-  | { view: "fields"; items: StudioField[] }
-  | { view: "clauses"; items: SlashClause[] };
-
-/** The new field's path for a "New field" row. Safe field paths such as
- *  `party.name` / `line-item` are kept verbatim; unsafe or invalid queries fall
- *  back to the sanitizer (`slugify` would strip the `.`/`-` and mangle the name
- *  the user typed). */
-const createFieldPathFromQuery = (trimmed: string): string => {
-  if (trimmed === "") {
-    return "field";
-  }
-  if (isSafeFieldPath(trimmed)) {
-    return trimmed;
-  }
-  const sanitized = sanitizeFieldPath(trimmed);
-  return isSafeFieldPath(sanitized) ? sanitized : "field";
-};
-
-/** The first field path not already taken: `field`, then `field_2`, `field_3`…
- *  Shared by the blank `/` create row and the insert-field action so a generic
- *  field can always be added even when `field` already exists. */
+/** The first field path not already taken: `field`, then `field_2`, `field_3`… */
 const uniqueFieldPath = (base: string, fields: StudioField[]): string => {
   let path = base;
   for (let n = 2; fields.some((field) => field.path === path); n++) {
     path = `${base}_${n}`;
   }
   return path;
-};
-
-/** Build the root rows for the current query. The typed text both filters the
- *  rows and (for "New field") becomes the new field's name. */
-const buildSlashRootItems = (
-  query: string,
-  fields: StudioField[],
-): SlashRootItem[] => {
-  const trimmed = query.trim();
-  const needle = trimmed.toLowerCase();
-  const matches = (...keywords: string[]): boolean =>
-    keywords.some((kw) => kw.includes(needle));
-
-  // Empty query: browse the grouped entry points (create a field/condition, or
-  // open the existing-field / clause submenus). A blank create row offers a
-  // fresh unique path (field_2, …) so a generic field can always be added.
-  if (trimmed === "") {
-    const items: SlashRootItem[] = [
-      { kind: "create-field", path: uniqueFieldPath("field", fields) },
-      { kind: "create-condition" },
-    ];
-    if (fields.length > 0) {
-      items.push({ kind: "open-fields" });
-    }
-    items.push({ kind: "open-clauses" });
-    return items;
-  }
-
-  // Typed query: matching existing fields first (one keystroke/click to reuse),
-  // then keyword-matched commands (clause search, condition), then the generic
-  // create-new-field fallback last. Order matters: the highlight resets to row 0
-  // on each keystroke, so "/if" or "/clause" + Enter must run the command rather
-  // than create a field literally named "if"/"clause".
-  const createPath = createFieldPathFromQuery(trimmed);
-  const reuseExact = fields.some((field) => field.path === createPath);
-  const items: SlashRootItem[] = matchingSlashFields(query, fields).map(
-    (field) => ({
-      kind: "field" as const,
-      path: field.path,
-      label: field.label === "" ? field.path : field.label,
-    }),
-  );
-  if (matches("clause")) {
-    items.push({ kind: "open-clauses" });
-  }
-  if (matches("condition", "if")) {
-    items.push({ kind: "create-condition" });
-  }
-  if (!reuseExact) {
-    items.push({ kind: "create-field", path: createPath });
-  }
-  return items;
-};
-
-/** Fields offered in the "Existing field" submenu, filtered by the submenu
- *  query (the same typed `/` text). */
-const matchingSlashFields = (
-  query: string,
-  fields: StudioField[],
-): StudioField[] => {
-  const needle = query.trim().toLowerCase();
-  const matches = (field: StudioField): boolean => {
-    if (needle === "") {
-      return true;
-    }
-    const label = field.label === "" ? field.path : field.label;
-    return (
-      field.path.toLowerCase().includes(needle) ||
-      label.toLowerCase().includes(needle)
-    );
-  };
-  return fields.filter(matches).slice(0, SLASH_MENU_MAX_FIELDS);
-};
-
-/** Row count for the active view, so the key handler can wrap/clamp the
- *  highlight without rebuilding the clause list (clauses come from a query
- *  whose length the caller passes in). */
-const slashRowCount = (
-  view: SlashView,
-  query: string,
-  fields: StudioField[],
-  clauseCount: number,
-): number => {
-  if (view === "fields") {
-    return matchingSlashFields(query, fields).length;
-  }
-  if (view === "clauses") {
-    return clauseCount;
-  }
-  return buildSlashRootItems(query, fields).length;
-};
-
-/** Session-lived answers for the selection popover, keyed by exact selection
- *  + surrounding context + the known-paths list; bounded FIFO so it cannot
- *  grow unchecked. */
-const GESTURE_ENRICHMENT_CACHE_MAX = 100;
-const gestureEnrichmentCache = new BoundedMap<string, GestureEnrichment>(
-  GESTURE_ENRICHMENT_CACHE_MAX,
-);
-
-/** Progressive AI proposal for the popover's AI row. `fieldPath` carries the
- *  model's claim that the selection is another occurrence of an existing
- *  field; it is re-checked against the session's fields at render/accept
- *  time. */
-type GestureEnrichment =
-  | { status: "idle" }
-  | { status: "loading" }
-  | {
-      status: "ready";
-      label: string | undefined;
-      inputType: TemplateEditableField["inputType"] | undefined;
-      aiPrompt: string | undefined;
-      fieldPath: string | undefined;
-    };
-
-/** Tiny stable digest (a polynomial rolling hash over the joined list) so
- *  the enrichment cache key reflects the known-paths list without storing
- *  it verbatim. */
-const HASH_MODULUS = 2 ** 48;
-const hashStrings = (values: readonly string[]): string => {
-  let hash = 5381;
-  for (const char of values.join("\u0000")) {
-    hash = (hash * 31 + (char.codePointAt(0) ?? 0)) % HASH_MODULUS;
-  }
-  return hash.toString(36);
-};
-
-/** The session's field paths as they ride along in the enrichment
- *  instruction: capped in count and joined length. */
-const enrichmentKnownPaths = (fields: readonly StudioField[]): string[] => {
-  const paths: string[] = [];
-  let budget = GESTURE_ENRICH_PATHS_MAX_CHARS;
-  for (const field of fields.slice(0, GESTURE_ENRICH_MAX_PATHS)) {
-    budget -= field.path.length + 2;
-    if (budget < 0) {
-      break;
-    }
-    paths.push(field.path);
-  }
-  return paths;
 };
 
 // True when the caret sits inside an `{{#each}}…{{/each}}` body, paired by
@@ -1080,56 +796,6 @@ export const TemplateStudioPage = ({
     insertOrWrapBlock("{{#each items}}", "{{/each}}", "items", range, true);
   const insertClause = () => insertInline("{{@clause:Clause}}");
 
-  // A clause-slot name not already taken by another `{{@clause:...}}` marker
-  // in the document, or claimed by a deferred slot rename: a pending step's
-  // fromSlot is absent from the document but still on the server row (or
-  // transiently claimed mid-replay), so linking a new clause under it would
-  // collide with the per-template unique-slot constraint at PUT time.
-  const uniqueClauseSlotName = (base: string): string => {
-    const view = editorViewRef.current;
-    const seed = base === "" ? "clause" : base;
-    const taken = new Set([
-      ...(view
-        ? getTemplateDirectives(view.state).flatMap((d) =>
-            d.kind === "clause" ? [d.expr] : [],
-          )
-        : []),
-      ...useTemplateStudioStore
-        .getState()
-        .pendingSlotRenames.flatMap((step) => [step.fromSlot, step.slotName]),
-    ]);
-    let candidate = seed;
-    for (let n = 2; taken.has(candidate); n++) {
-      candidate = `${seed}_${n}`;
-    }
-    return candidate;
-  };
-
-  // Link a chosen library clause to the slot just inserted, reusing the same
-  // template-clause link endpoint the clause inspector + link dialog use.
-  const linkClauseToSlot = async (clauseId: string, slotName: string) => {
-    const response = await api
-      .templates({ templateId: toSafeId<"template">(templateId) })
-      .clauses.put({ clauseId: toSafeId<"clause">(clauseId), slotName });
-    if (response.error) {
-      stellaToast.add({
-        type: "error",
-        title: t("clauses.linkFailed"),
-        description: userErrorMessage(
-          response.error,
-          t("common.unexpectedError"),
-        ),
-      });
-      return;
-    }
-    void queryClient.invalidateQueries({
-      queryKey: knowledgeKeys.templates.clauses(
-        activeOrganizationId,
-        templateId,
-      ),
-    });
-  };
-
   /** Explicit-click block mirror: wrap the parallel cell's paragraphs in
    *  the same block. The opener's live expression is read at click time via
    *  the synced selected directive (the user typically renames the
@@ -1189,286 +855,22 @@ export const TemplateStudioPage = ({
     }
   };
 
-  // ── Selection → gesture popover ──────────────────────────
-  // Selecting plain prose floats the structural gestures next to the
-  // selection: Make field leads, wrap-in-condition / wrap-in-loop follow.
-  // After the selection has been stable for a moment, the model proposes a
-  // configuration for the would-be field and the Make-field row enriches;
-  // the buttons themselves never wait for it.
   const overlayHostRef = useRef<HTMLDivElement | null>(null);
-  const [gesture, setGestureState] = useState<SelectionGesture | null>(null);
-  // Mirror for handlers that fire outside the render cycle (debounced
-  // callbacks, in-flight enrichment) and need the freshest value.
-  const gestureRef = useRef<SelectionGesture | null>(null);
-  const setGesture = (next: SelectionGesture | null) => {
-    gestureRef.current = next;
-    setGestureState(next);
-  };
-  const [enrichment, setEnrichment] = useState<GestureEnrichment>({
-    status: "idle",
-  });
-  // Bumped on every hide/new fetch so stale enrichment responses are dropped.
-  const enrichSeqRef = useRef(0);
-
-  // useCallback (not React Compiler memoization) because the dismiss-listener
-  // effect depends on it.
-  const hideGesture = useCallback(() => {
-    enrichSeqRef.current++;
-    gestureRef.current = null;
-    setGestureState(null);
-    setEnrichment({ status: "idle" });
-  }, []);
-
-  const showGesture = useDebouncedCallback((sel: GestureSelection) => {
-    const view = editorViewRef.current;
-    const host = overlayHostRef.current;
-    if (!view || !host) {
-      return;
-    }
-    const { from, to } = view.state.selection;
-    if (from === to || from !== sel.from || to !== sel.to) {
-      return;
-    }
-    // Only for selections made in the document itself: while the AI chat bar
-    // (or any surface outside the editor) holds focus, stay away.
-    const scrollContainer = view.dom.closest("[data-folio-scroll]");
-    if (!scrollContainer || !scrollContainer.contains(document.activeElement)) {
-      return;
-    }
-    // Existing {{markers}} have their own inspector faces; the gesture
-    // popover is only for plain prose.
-    const overlapsDirective = getTemplateDirectives(view.state).some(
-      (range) => range.from < to && range.to > from,
-    );
-    if (overlapsDirective) {
-      return;
-    }
-    // Folio paints the selection highlights asynchronously after the
-    // selection change; retry across a few frames until they exist.
-    let attempts = 0;
-    const read = () => {
-      const liveView = editorViewRef.current;
-      if (!liveView) {
-        return;
-      }
-      const live = liveView.state.selection;
-      if (live.from !== sel.from || live.to !== sel.to) {
-        return;
-      }
-      const rect = getFolioSelectionViewportRect(liveView);
-      if (!rect) {
-        attempts++;
-        if (attempts < 10) {
-          requestAnimationFrame(read);
-        }
-        return;
-      }
-      const hostRect = host.getBoundingClientRect();
-      const fitsBelow =
-        rect.bottom -
-          hostRect.top +
-          GESTURE_POPOVER_OFFSET_PX +
-          GESTURE_POPOVER_EST_HEIGHT_PX <
-        host.clientHeight;
-      const placement = fitsBelow ? "below" : "above";
-      const center = rect.left + rect.width / 2 - hostRect.left;
-      // The popover is also capped at the host width (max-w: …,100%), so in a
-      // pane narrower than its natural width the real half-width is host/2 —
-      // clamp with that to keep it from spilling past the edge.
-      const halfWidth = Math.min(
-        GESTURE_POPOVER_HALF_WIDTH_PX,
-        host.clientWidth / 2,
-      );
-      setGesture({
-        ...sel,
-        left: Math.min(
-          Math.max(center, halfWidth),
-          Math.max(halfWidth, host.clientWidth - halfWidth),
-        ),
-        top:
-          placement === "below"
-            ? rect.bottom - hostRect.top
-            : rect.top - hostRect.top,
-        placement,
-      });
-    };
-    requestAnimationFrame(read);
-  }, GESTURE_SHOW_DELAY_MS);
-
-  const fetchGestureSuggestion = async (sel: GestureSelection, seq: number) => {
-    const view = editorViewRef.current;
-    if (!view) {
-      return;
-    }
-    const contextText = paragraphsAroundSelection(view.state, sel.from);
-    const knownPaths = enrichmentKnownPaths(
-      useTemplateStudioStore.getState().fields,
-    );
-    // Re-selecting the same text in the same context is common (the popover
-    // dismisses on any click); serve the model's previous answer instead of
-    // paying for it again. The known-paths list changes rarely; its hash
-    // keys the existing-field part of the answer.
-    const cacheKey = `${sel.text}\u0000${contextText}\u0000${hashStrings(knownPaths)}`;
-    const cached = gestureEnrichmentCache.get(cacheKey);
-    if (cached !== undefined) {
-      setEnrichment(cached);
-      return;
-    }
-    setEnrichment({ status: "loading" });
-    const existingFieldsClause =
-      knownPaths.length === 0
-        ? ""
-        : ` The template already has these fields: ${knownPaths.join(", ")}.` +
-          ` If the selection is just another occurrence of one of them, set` +
-          ` fieldPath to that exact existing path instead of inventing a` +
-          ` new field.`;
-    const response = await api.templates["suggest-fields"].post({
-      text: contextText,
-      instructions:
-        `The user selected this exact text: "${sel.text}". Propose exactly ` +
-        `ONE field for that exact selection (literalText must equal it): ` +
-        `its label and input type, plus an aiPrompt only when the selection ` +
-        `is free-form prose that AI should draft at fill ` +
-        `time.${existingFieldsClause}`,
-    });
-    if (seq !== enrichSeqRef.current) {
-      return;
-    }
-    const match = response.error ? null : response.data.suggestions.at(0);
-    if (!match) {
-      setEnrichment({ status: "idle" });
-      return;
-    }
-    const ready: GestureEnrichment = {
-      status: "ready",
-      label: match.label,
-      inputType:
-        match.inputType !== undefined && isInputType(match.inputType)
-          ? match.inputType
-          : undefined,
-      aiPrompt: match.aiPrompt,
-      fieldPath: match.fieldPath,
-    };
-    gestureEnrichmentCache.set(cacheKey, ready);
-    setEnrichment(ready);
-  };
-
-  const enrichGesture = useDebouncedCallback((sel: GestureSelection) => {
-    const shown = gestureRef.current;
-    if (shown === null || shown.from !== sel.from || shown.to !== sel.to) {
-      return;
-    }
-    const seq = ++enrichSeqRef.current;
-    fetchGestureSuggestion(sel, seq).catch((error: unknown) => {
-      if (seq !== enrichSeqRef.current) {
-        return;
-      }
-      getAnalytics().captureError(error);
-      setEnrichment({ status: "idle" });
-      stellaToast.add({ title: t("errors.actionFailed"), type: "error" });
-    });
-  }, GESTURE_ENRICH_DELAY_MS);
-
-  const onGestureSelectionChange = (sel: GestureSelection) => {
-    if (sel.text.trim() === "") {
-      showGesture.cancel();
-      enrichGesture.cancel();
-      hideGesture();
-      return;
-    }
-    const shown = gestureRef.current;
-    if (shown !== null && (shown.from !== sel.from || shown.to !== sel.to)) {
-      hideGesture();
-    }
-    showGesture(sel);
-    enrichGesture(sel);
-  };
-
-  const dismissGesture = () => {
-    showGesture.cancel();
-    enrichGesture.cancel();
-    hideGesture();
-  };
-
-  /** The popover's "Existing field…" list: reuse a field over the captured
-   *  selection. */
-  const applyExistingFieldGesture = (path: string) => {
-    const shown = gestureRef.current;
-    if (shown === null) {
-      return;
-    }
-    insertExistingFieldAt(path, {
-      range: { from: shown.from, to: shown.to },
-    });
-    dismissGesture();
-  };
-
-  /** Wrap the captured selection in an existing condition (reuse), referencing
-   *  it by name so the gate shares the condition-field's single source. */
-  const applyExistingConditionGesture = (conditionName: string) => {
-    const shown = gestureRef.current;
-    if (shown === null) {
-      return;
-    }
-    insertOrWrapBlock(
-      `{{#if ${conditionName}}}`,
-      "{{/if}}",
-      conditionName,
-      { from: shown.from, to: shown.to },
-      true,
-    );
-    dismissGesture();
-  };
-
-  /** Accept the AI proposal row: when the model recognized the selection as
-   *  another occurrence of an existing field, reuse that field; otherwise
-   *  make a new field carrying the proposed configuration. */
-  const applyAiGesture = () => {
-    const shown = gestureRef.current;
-    if (shown === null || enrichment.status !== "ready") {
-      return;
-    }
-    const range = { from: shown.from, to: shown.to };
-    const existing =
-      enrichment.fieldPath === undefined
-        ? undefined
-        : useTemplateStudioStore
-            .getState()
-            .fields.find((f) => f.path === enrichment.fieldPath);
-    if (existing !== undefined) {
-      insertExistingFieldAt(existing.path, { range });
-    } else {
-      const path = makeField(range);
-      if (path !== null) {
-        upsertField(path, {
-          ...(enrichment.label !== undefined && enrichment.label !== ""
-            ? { label: enrichment.label }
-            : {}),
-          ...(enrichment.inputType !== undefined
-            ? { inputType: enrichment.inputType }
-            : {}),
-          ...(enrichment.aiPrompt !== undefined
-            ? { aiPrompt: enrichment.aiPrompt }
-            : {}),
-        });
-      }
-    }
-    dismissGesture();
-  };
-
-  // Each button acts on the selection captured when the popover anchored, so
-  // a click can never target a drifted live selection.
-  const applyGesture = (kind: "field" | "if" | "each" | "clause") => {
-    const shown = gestureRef.current;
-    if (shown === null) {
-      return;
-    }
-    const range = { from: shown.from, to: shown.to };
-    if (kind === "field") {
-      // The plain row is deliberately generic; the AI proposal accepts via
-      // its own row below the separator.
-      makeField(range);
-    } else if (kind === "clause") {
+  const selectionGesture = useTemplateStudioSelectionGesture({
+    editorViewRef,
+    overlayHostRef,
+    makeField,
+    insertExistingField: (path, range) =>
+      insertExistingFieldAt(path, { range }),
+    insertExistingCondition: (conditionName, range) =>
+      insertOrWrapBlock(
+        `{{#if ${conditionName}}}`,
+        "{{/if}}",
+        conditionName,
+        range,
+        true,
+      ),
+    insertClause: (range) =>
       withEditorView((view) => {
         const slotName = slugify(
           view.state.doc.textBetween(range.from, range.to, " "),
@@ -1480,498 +882,19 @@ export const TemplateStudioPage = ({
         );
         view.focus();
         markDirty();
-      });
-    } else if (kind === "if") {
-      wrapBlockWithMirrorOffer("if", range);
-    } else {
-      wrapBlockWithMirrorOffer("each", range);
-    }
-    dismissGesture();
-  };
-
-  // Escape, any scroll, and the (custom) context menu all dismiss the
-  // popover: the selection survives, only the floating affordance leaves.
-  const gestureShown = gesture !== null;
-  useExternalSyncEffect(() => {
-    if (!gestureShown) {
-      return undefined;
-    }
-    const host = overlayHostRef.current;
-    const onKeyDown = (event: KeyboardEvent) => {
-      if (event.key === "Escape") {
-        hideGesture();
-      }
-    };
-    const dismiss = () => hideGesture();
-    window.addEventListener("keydown", onKeyDown);
-    host?.addEventListener("scroll", dismiss, {
-      capture: true,
-      passive: true,
-    });
-    host?.addEventListener("contextmenu", dismiss, { capture: true });
-    return () => {
-      window.removeEventListener("keydown", onKeyDown);
-      host?.removeEventListener("scroll", dismiss, {
-        capture: true,
-      });
-      host?.removeEventListener("contextmenu", dismiss, { capture: true });
-    };
-  }, [gestureShown, hideGesture]);
-
-  useExternalSyncEffect(
-    () => () => {
-      showGesture.cancel();
-      enrichGesture.cancel();
-    },
-    [showGesture, enrichGesture],
-  );
-
-  // ── Slash-command menu ───────────────────────────────────
-  // Typing `/` at a marker boundary opens a keyboard-first, command-palette-style
-  // menu at the caret. The root lists two direct creators (New field, Condition)
-  // and two submenu openers (Existing field, Clause); the typed text filters the
-  // rows and doubles as a new field's name. Folio's plugin owns the trigger and
-  // forwards navigation keys; this host renders the menu, tracks the highlighted
-  // row + active submenu, and performs the insertion.
-  const [slash, setSlashState] = useState<SlashMenu | null>(null);
-  const [slashView, setSlashViewState] = useState<SlashView>("root");
-  const [slashHighlight, setSlashHighlightState] = useState(0);
-  // Mirrors for the plugin's synchronous key handler, which fires outside the
-  // React render cycle and needs the freshest view + highlight + row count.
-  const slashViewRef = useRef<SlashView>("root");
-  const slashHighlightRef = useRef(0);
-  // The `/` anchor the menu is currently positioned for; a change means a fresh
-  // trigger that needs (re)positioning, vs. the same trigger's query growing.
-  const slashFromRef = useRef<number | null>(null);
-  const setSlashView = (view: SlashView) => {
-    slashViewRef.current = view;
-    setSlashViewState(view);
-  };
-  const setSlashHighlight = (index: number) => {
-    slashHighlightRef.current = index;
-    setSlashHighlightState(index);
-  };
-  const studioFields = useTemplateStudioStore((s) => s.fields);
-
-  // Clause library for the clause submenu, searched by the live `/` query.
-  // Enabled only while that submenu is open so the root menu costs nothing.
-  const slashClausesEnabled = slash !== null && slashView === "clauses";
-  // Debounced so search-as-you-type filters server-side without a query per
-  // keystroke. The server does prefix full-text matching (scales with the
-  // library); the client never holds the whole library.
-  const [debouncedClauseSearch] = useDebounce(
-    slashView === "clauses" ? (slash?.query ?? "") : "",
-    120,
-  );
-  const { data: slashClauseData } = useQuery({
-    ...clauseLibraryOptions(activeOrganizationId, {
-      search: debouncedClauseSearch,
-      limit: SLASH_MENU_CLAUSE_LIMIT,
-    }),
-    enabled: slashClausesEnabled,
+      }),
+    wrapBlock: wrapBlockWithMirrorOffer,
+    upsertField,
   });
-  const slashClauses: SlashClause[] = useMemo(
-    () =>
-      slashClauseData && "items" in slashClauseData
-        ? slashClauseData.items
-        : [],
-    [slashClauseData],
-  );
-  // Effect-synced (not render-mirrored) so the synchronous key handler reads the
-  // freshest clause list — and the search it was fetched for — without assigning
-  // a ref during render.
-  const slashClausesRef = useRef<SlashClause[]>(slashClauses);
-  const debouncedClauseSearchRef = useRef(debouncedClauseSearch);
-  useExternalSyncEffect(() => {
-    slashClausesRef.current = slashClauses;
-    debouncedClauseSearchRef.current = debouncedClauseSearch;
-  }, [slashClauses, debouncedClauseSearch]);
-
-  // The rows visible for the current view + query, as a discriminated union so
-  // render and the key handler narrow without unsafe casts.
-  const slashRows = useMemo((): SlashRows => {
-    if (slash === null) {
-      return { view: "root", items: [] };
-    }
-    if (slashView === "fields") {
-      return {
-        view: "fields",
-        items: matchingSlashFields(slash.query, studioFields),
-      };
-    }
-    if (slashView === "clauses") {
-      return { view: "clauses", items: slashClauses };
-    }
-    return {
-      view: "root",
-      items: buildSlashRootItems(slash.query, studioFields),
-    };
-  }, [slash, slashView, studioFields, slashClauses]);
-
-  // Re-anchor the menu at the caret. Geometry only: the query is updated
-  // synchronously in `onSlashMenuChange`, decoupled from the caret rect, so
-  // filtering never waits on (or is dropped by) the paged editor's async caret
-  // repaint. Folio paints the caret a frame or two after the selection change,
-  // so retry briefly until it exists, then merge left/top/placement into the
-  // already-rendered menu without disturbing the live query.
-  const positionSlashMenu = useCallback((from: number) => {
-    const host = overlayHostRef.current;
-    if (!host) {
-      return;
-    }
-    let attempts = 0;
-    const read = () => {
-      const liveView = editorViewRef.current;
-      const live = liveView ? getTemplateSlashMenu(liveView.state) : null;
-      // Bail if the trigger is gone or has been replaced by a newer one (a
-      // later keystroke re-anchored at a different `/`); the matching call will
-      // position that one.
-      if (
-        liveView === null ||
-        live === null ||
-        !live.active ||
-        live.from !== from
-      ) {
-        return;
-      }
-      const rect = getFolioCaretViewportRect(liveView);
-      if (!rect) {
-        attempts++;
-        if (attempts < 10) {
-          // eslint-disable-next-line react/react-compiler -- recursive local function flagged as its own dependency; `read` is not a reactive value and cannot be added to the useCallback deps
-          requestAnimationFrame(read);
-        }
-        return;
-      }
-      const hostRect = host.getBoundingClientRect();
-      const caretLeft = rect.left - hostRect.left;
-      const caretTop = rect.top - hostRect.top;
-      const caretBottom = rect.bottom - hostRect.top;
-      // Prefer opening ABOVE the caret (below-placement collides with the
-      // suggested-action chips and the floating AI bar pinned at the bottom),
-      // but only when there is actually room above — otherwise the menu would
-      // render off the top of the editor. Near the top, fall back to below.
-      const fitsAbove =
-        caretTop - SLASH_MENU_OFFSET_PX - SLASH_MENU_EST_HEIGHT_PX >= 0;
-      const placement = fitsAbove ? "above" : "below";
-      // Anchor the menu's start edge at the caret, clamping against the ACTUAL
-      // rendered width so it never spills off the inline-end edge. Below the `sm`
-      // breakpoint `MenuPreviewLayout` hides the preview pane, so the menu is just
-      // the list column — clamping against the full desktop width there would
-      // shove a near-caret menu needlessly to the left.
-      const renderedWidth =
-        window.innerWidth >= SLASH_MENU_PREVIEW_BREAKPOINT_PX
-          ? SLASH_MENU_WIDTH_PX
-          : SLASH_MENU_LIST_WIDTH_PX;
-      const left = Math.max(
-        0,
-        Math.min(caretLeft, host.clientWidth - renderedWidth),
-      );
-      setSlashState((prev) =>
-        prev === null
-          ? prev
-          : {
-              ...prev,
-              left,
-              top: placement === "below" ? caretBottom : caretTop,
-              placement,
-            },
-      );
-    };
-    requestAnimationFrame(read);
-  }, []);
-
-  const onSlashMenuChange = (state: TemplateSlashMenuState) => {
-    if (!state.active) {
-      setSlashState(null);
-      setSlashView("root");
-      slashFromRef.current = null;
-      return;
-    }
-    // Update the query SYNCHRONOUSLY so the visible rows filter on every
-    // keystroke. The menu is anchored at the `/` and does NOT follow the caret
-    // as the query grows, so geometry is computed once on open; default
-    // placement is "below" so a not-yet-positioned menu shows on-screen below
-    // the caret rather than translating its height off the top of the editor.
-    const isNewTrigger = slashFromRef.current !== state.from;
-    slashFromRef.current = state.from;
-    setSlashState((prev) => ({
-      from: state.from,
-      query: state.query,
-      left: prev?.left ?? 0,
-      top: prev?.top ?? 0,
-      placement: prev?.placement ?? "below",
-    }));
-    // Reset the highlight to the top row whenever the query changes; the row
-    // list is rebuilt and the previous index may point at a gone row.
-    setSlashHighlight(0);
-    // Position only when the trigger first anchors at a new `/`: the menu stays
-    // pinned to the `/` as the query grows, and re-reading the caret rect per
-    // keystroke both churns and races the paged-editor relayout.
-    if (isNewTrigger) {
-      positionSlashMenu(state.from);
-    }
-  };
-
-  const dismissSlash = useCallback(() => {
-    const view = editorViewRef.current;
-    if (view && getTemplateSlashMenu(view.state).active) {
-      view.dispatch(clearTemplateSlashMenu(view.state.tr));
-    }
-    setSlashState(null);
-    // eslint-disable-next-line react/react-compiler -- `setSlashView` closes only over the stable setSlashViewState setter and slashViewRef, so a stale closure is a no-op; empty deps are intentional to keep dismissSlash referentially stable for the [slashShown, dismissSlash] effect below
-    setSlashView("root");
-    slashFromRef.current = null;
-  }, []);
-
-  // Insert a fresh field marker, selecting the name so syncSelection opens the
-  // field face. Reuses `from` from the consumed `/query` range.
-  const slashInsertNewField = (
-    view: EditorView,
-    consumed: { tr: Transaction; from: number },
-    path: string,
-  ) => {
-    const tr = consumed.tr.insertText(`{{${path}}}`, consumed.from);
-    const namePos = consumed.from + 2;
-    tr.setSelection(
-      TextSelection.create(tr.doc, namePos, namePos + path.length),
-    ).scrollIntoView();
-    view.dispatch(tr);
-    view.focus();
-    upsertField(path, {});
-    markDirty();
-  };
-
-  // Switch menu level. Reset the typed `/` query to blank so each level filters
-  // from scratch (the submenu's search is its own; the root's is its own).
-  const enterSlashSubmenu = (next: SlashView) => {
-    const editor = editorViewRef.current;
-    if (editor) {
-      const tr = resetTemplateSlashQuery(editor.state);
-      if (tr !== null) {
-        editor.dispatch(tr);
-      }
-    }
-    setSlashView(next);
-    setSlashHighlight(0);
-  };
-
-  // Activate a root row: direct creators insert immediately; submenu openers
-  // switch the view instead.
-  const activateSlashRoot = (item: SlashRootItem) => {
-    if (item.kind === "open-fields") {
-      enterSlashSubmenu("fields");
-      return;
-    }
-    if (item.kind === "open-clauses") {
-      enterSlashSubmenu("clauses");
-      return;
-    }
-    const view = editorViewRef.current;
-    if (!view) {
-      return;
-    }
-    const consumed = consumeTemplateSlashQuery(view.state);
-    if (consumed === null) {
-      return;
-    }
-    if (item.kind === "create-field") {
-      slashInsertNewField(view, consumed, item.path);
-      dismissSlash();
-      return;
-    }
-    if (item.kind === "field") {
-      // Reuse an existing field picked inline from the filtered root list.
-      view.dispatch(
-        consumed.tr
-          .insertText(`{{${item.path}}}`, consumed.from)
-          .scrollIntoView(),
-      );
-      view.focus();
-      markDirty();
-      dismissSlash();
-      return;
-    }
-    // create-condition: drop the query, then insert the block at the collapsed
-    // caret via the shared block helper.
-    view.dispatch(consumed.tr.scrollIntoView());
-    insertCondition();
-    dismissSlash();
-  };
-
-  // Reuse a registered field at the caret without forcing focus there.
-  const activateSlashField = (path: string) => {
-    const view = editorViewRef.current;
-    if (!view) {
-      return;
-    }
-    const consumed = consumeTemplateSlashQuery(view.state);
-    if (consumed === null) {
-      return;
-    }
-    view.dispatch(
-      consumed.tr.insertText(`{{${path}}}`, consumed.from).scrollIntoView(),
-    );
-    view.focus();
-    markDirty();
-    dismissSlash();
-  };
-
-  // Insert a clause slot named after the chosen clause and link that slot to it,
-  // reusing the same slot + link mechanism as the clause inspector.
-  const activateSlashClause = (clause: { id: string; title: string }) => {
-    const view = editorViewRef.current;
-    if (!view) {
-      return;
-    }
-    // Both the Enter and click paths land here. Don't insert a stale clause: if
-    // the 120ms-debounced clause search has not caught up to the live query, the
-    // visible rows are from the previous query — bail until they refresh.
-    const live = getTemplateSlashMenu(view.state);
-    if (live.active && debouncedClauseSearchRef.current !== live.query) {
-      return;
-    }
-    const consumed = consumeTemplateSlashQuery(view.state);
-    if (consumed === null) {
-      return;
-    }
-    const slotName = uniqueClauseSlotName(slugify(clause.title));
-    view.dispatch(
-      consumed.tr
-        .insertText(`{{@clause:${slotName}}}`, consumed.from)
-        .scrollIntoView(),
-    );
-    view.focus();
-    markDirty();
-    dismissSlash();
-    void linkClauseToSlot(clause.id, slotName);
-  };
-
-  const onSlashMenuKeyAction = (
-    action: TemplateSlashMenuKeyAction,
-  ): boolean => {
-    // Derive everything fresh from the live trigger + current view so the
-    // synchronous key handler never acts on a stale render.
-    const view = editorViewRef.current;
-    const live = view ? getTemplateSlashMenu(view.state) : null;
-    if (live === null || !live.active) {
-      return false;
-    }
-    const currentView = slashViewRef.current;
-    if (action === "back") {
-      if (currentView === "root") {
-        return false;
-      }
-      enterSlashSubmenu("root");
-      return true;
-    }
-    if (action === "dismiss") {
-      if (currentView === "root") {
-        return false;
-      }
-      enterSlashSubmenu("root");
-      return true;
-    }
-    const rowCount = slashRowCount(
-      currentView,
-      live.query,
-      useTemplateStudioStore.getState().fields,
-      slashClausesRef.current.length,
-    );
-    if (rowCount === 0) {
-      // Keep arrows/enter swallowed so a momentarily empty submenu does not move
-      // the document caret out from under the open menu.
-      return action === "up" || action === "down" || action === "commit";
-    }
-    if (action === "up") {
-      setSlashHighlight((slashHighlightRef.current - 1 + rowCount) % rowCount);
-      return true;
-    }
-    if (action === "down") {
-      setSlashHighlight((slashHighlightRef.current + 1) % rowCount);
-      return true;
-    }
-    const index = Math.min(slashHighlightRef.current, rowCount - 1);
-    if (currentView === "root") {
-      const items = buildSlashRootItems(
-        live.query,
-        useTemplateStudioStore.getState().fields,
-      );
-      const item = items.at(index);
-      if (item === undefined) {
-        return false;
-      }
-      // ArrowRight only enters submenu openers; on a creator row it is a no-op.
-      if (action === "forward") {
-        if (item.kind === "open-fields" || item.kind === "open-clauses") {
-          activateSlashRoot(item);
-          return true;
-        }
-        return false;
-      }
-      activateSlashRoot(item);
-      return true;
-    }
-    // forward inside a submenu has no deeper level.
-    if (action === "forward") {
-      return false;
-    }
-    if (currentView === "fields") {
-      const field = matchingSlashFields(
-        live.query,
-        useTemplateStudioStore.getState().fields,
-      ).at(index);
-      if (field === undefined) {
-        return false;
-      }
-      activateSlashField(field.path);
-      return true;
-    }
-    // Enter on a clause: `activateSlashClause` bails if the debounced search is
-    // stale, so a too-fast Enter waits for refreshed rows instead of inserting
-    // the wrong clause. Swallow the key either way so it never adds a newline.
-    const clause = slashClausesRef.current.at(index);
-    if (clause === undefined) {
-      return false;
-    }
-    activateSlashClause(clause);
-    return true;
-  };
-
-  // A USER scroll of the document (wheel/touch) — or a context action — tears the
-  // menu down, since it is anchored to the `/` and would otherwise float. We
-  // listen for wheel/touchmove, NOT the generic `scroll` event: typing near the
-  // bottom programmatically scrolls the editor to keep the caret in view, and
-  // that must not close the menu mid-filter. Scrolling within the menu's own list
-  // is excluded so the user can scroll a long result set.
-  const slashShown = slash !== null;
-  useExternalSyncEffect(() => {
-    if (!slashShown) {
-      return undefined;
-    }
-    const host = overlayHostRef.current;
-    const dismiss = (event: Event) => {
-      if (
-        event.target instanceof Element &&
-        event.target.closest('[role="listbox"]')
-      ) {
-        return;
-      }
-      dismissSlash();
-    };
-    host?.addEventListener("wheel", dismiss, { capture: true, passive: true });
-    host?.addEventListener("touchmove", dismiss, {
-      capture: true,
-      passive: true,
-    });
-    host?.addEventListener("contextmenu", dismiss, { capture: true });
-    return () => {
-      host?.removeEventListener("wheel", dismiss, { capture: true });
-      host?.removeEventListener("touchmove", dismiss, { capture: true });
-      host?.removeEventListener("contextmenu", dismiss, { capture: true });
-    };
-  }, [slashShown, dismissSlash]);
-
+  const slashMenu = useTemplateStudioSlashMenu({
+    activeOrganizationId,
+    templateId,
+    editorViewRef,
+    overlayHostRef,
+    insertCondition,
+    markDirty,
+    upsertField,
+  });
   // Loop recipes mirror insertOrWrapBlock's caret branch: the opener, one
   // marker paragraph per field, and the closer land after the current
   // paragraph (block directives must occupy their own paragraph).
@@ -2740,38 +1663,18 @@ export const TemplateStudioPage = ({
                 setHasSelection(state?.hasSelection ?? false);
                 syncSelection();
               }}
-              onSelectionTextChange={onGestureSelectionChange}
-              onSlashMenuChange={onSlashMenuChange}
-              onSlashMenuKeyAction={onSlashMenuKeyAction}
+              onSelectionTextChange={selectionGesture.onSelectionTextChange}
+              onSlashMenuChange={slashMenu.onSlashMenuChange}
+              onSlashMenuKeyAction={slashMenu.onSlashMenuKeyAction}
               showTemplateDirectives={showDirectives}
             />
           </Suspense>
         </div>
-        {slash !== null && (
-          <SlashMenuPopover
-            slash={slash}
-            rows={slashRows}
-            highlight={slashHighlight}
-            fields={studioFields}
-            onHighlight={setSlashHighlight}
-            onActivateRoot={activateSlashRoot}
-            onActivateField={activateSlashField}
-            onActivateClause={activateSlashClause}
-            onBack={() => enterSlashSubmenu("root")}
-          />
+        {slashMenu.renderState !== null && (
+          <TemplateStudioSlashMenu {...slashMenu.renderState} />
         )}
-        {gesture !== null && (
-          <SelectionGesturePopover
-            enrichment={enrichment}
-            gesture={gesture}
-            onAcceptAi={applyAiGesture}
-            onInsertExisting={applyExistingFieldGesture}
-            onMakeClause={() => applyGesture("clause")}
-            onMakeField={() => applyGesture("field")}
-            onWrapEach={() => applyGesture("each")}
-            onWrapIf={() => applyGesture("if")}
-            onWrapIfExisting={applyExistingConditionGesture}
-          />
+        {selectionGesture.renderState !== null && (
+          <TemplateStudioSelectionGesture {...selectionGesture.renderState} />
         )}
         <TemplateStudioChat
           editorRef={editorRef}
@@ -2793,1014 +1696,6 @@ export const TemplateStudioPage = ({
 // inspector. The page (above) owns the document + actions and seeds the shared
 // session store this view reads from. `close-on-route-leave` is a backstop; the
 // page also closes the tab on unmount.
-/** The selection's containing paragraph plus one paragraph on each side —
- *  enough context for the model to read how a new field is used, read from
- *  the live doc because the field marker does not exist yet. */
-const paragraphsAroundSelection = (
-  state: EditorState,
-  from: number,
-): string => {
-  const $from = state.doc.resolve(from);
-  if ($from.depth === 0) {
-    return $from.parent.textContent.slice(0, 4000);
-  }
-  const container = $from.node($from.depth - 1);
-  const index = $from.index($from.depth - 1);
-  const texts: string[] = [];
-  for (const i of [index - 1, index, index + 1]) {
-    if (i < 0 || i >= container.childCount) {
-      continue;
-    }
-    const text = container.child(i).textContent;
-    if (text.trim() !== "") {
-      texts.push(text);
-    }
-  }
-  return texts.join("\n");
-};
-
-/** Popover buttons act on the captured selection; preventing mousedown keeps
- *  focus (and the painted selection) in the editor while clicking. */
-const keepEditorFocus = (event: { preventDefault: () => void }) => {
-  event.preventDefault();
-};
-
-/** Floating gesture menu anchored under (or above) the painted selection.
- *  Instant rows first (plain Make field, the existing-field list, the
- *  structural wraps); the AI proposal lands as its own row below a
- *  separator once the model answers. */
-const SelectionGesturePopover = ({
-  gesture,
-  enrichment,
-  onMakeField,
-  onInsertExisting,
-  onAcceptAi,
-  onWrapIf,
-  onWrapIfExisting,
-  onWrapEach,
-  onMakeClause,
-}: {
-  gesture: SelectionGesture;
-  enrichment: GestureEnrichment;
-  onMakeField: () => void;
-  onInsertExisting: (path: string) => void;
-  onAcceptAi: () => void;
-  onWrapIf: () => void;
-  onWrapIfExisting: (name: string) => void;
-  onWrapEach: () => void;
-  onMakeClause: () => void;
-}) => {
-  const t = useTranslations();
-  const fields = useTemplateStudioStore((s) => s.fields);
-  const [preview, setPreview] = useState<GestureInsertKind | null>(null);
-  return (
-    <div
-      className="bg-popover text-popover-foreground absolute z-50 flex max-h-[min(26rem,80vh)] max-w-[min(92vw,30rem,100%)] flex-col overflow-y-auto rounded-lg border p-1 shadow-lg/5 transition-opacity duration-100 starting:opacity-0"
-      role="group"
-      style={{
-        left: gesture.left,
-        top: gesture.top,
-        transform:
-          gesture.placement === "above"
-            ? `translate(-50%, calc(-100% - ${GESTURE_POPOVER_OFFSET_PX}px))`
-            : `translate(-50%, ${GESTURE_POPOVER_OFFSET_PX}px)`,
-      }}
-    >
-      <MenuPreviewLayout
-        preview={
-          <PreviewPane>
-            {preview ? (
-              <GestureInsertPreview key={preview} kind={preview} />
-            ) : (
-              <div className="text-muted-foreground flex h-full items-center justify-center text-center text-xs text-balance">
-                {t("templates.studio.previewHint")}
-              </div>
-            )}
-          </PreviewPane>
-        }
-      >
-        <GestureSplitRow
-          existing={fields.map((field) => ({
-            key: field.path,
-            icon: VALUE_TYPE_META[inputTypeValueKind(field.inputType)].icon,
-            label: field.label === "" ? field.path : field.label,
-            sublabel: field.label === "" ? undefined : field.path,
-            onSelect: () => onInsertExisting(field.path),
-          }))}
-          icon={BracesIcon}
-          label={t("templates.studio.makeField")}
-          onDefault={onMakeField}
-          onHighlight={() => setPreview("field")}
-          reuseLabel={t("templates.studio.existingField")}
-        />
-        <GestureSplitRow
-          existing={reusableConditions(fields, (key) => t(key)).map(
-            (condition) => ({
-              key: condition.ref,
-              icon: SplitIcon,
-              label: condition.label,
-              onSelect: () => onWrapIfExisting(condition.ref),
-            }),
-          )}
-          icon={SplitIcon}
-          label={t("templates.studio.showOnlyIf")}
-          onDefault={onWrapIf}
-          onHighlight={() => setPreview("if")}
-          reuseLabel={t("templates.studio.existingCondition")}
-        />
-        <Button
-          className="justify-start gap-2 font-normal"
-          onClick={onWrapEach}
-          onFocus={() => setPreview("each")}
-          onMouseDown={keepEditorFocus}
-          onMouseEnter={() => setPreview("each")}
-          size="sm"
-          variant="ghost"
-        >
-          <RepeatIcon className="text-muted-foreground size-3.5 shrink-0" />
-          {t("templates.studio.repeatForEach")}
-        </Button>
-        <Button
-          className="justify-start gap-2 font-normal"
-          onClick={onMakeClause}
-          onFocus={() => setPreview("clause")}
-          onMouseDown={keepEditorFocus}
-          onMouseEnter={() => setPreview("clause")}
-          size="sm"
-          variant="ghost"
-        >
-          <TextQuoteIcon className="text-muted-foreground size-3.5 shrink-0" />
-          {t("templates.studio.scopeClause")}
-        </Button>
-        {enrichment.status !== "idle" && (
-          <>
-            <Separator className="my-1" />
-            <GestureAiRow
-              enrichment={enrichment}
-              fields={fields}
-              onAcceptAi={onAcceptAi}
-            />
-          </>
-        )}
-      </MenuPreviewLayout>
-    </div>
-  );
-};
-
-/** A row's icon + translated label + optional right-aligned marker hint. */
-type SlashRowFace = {
-  icon: LucideIcon;
-  label: string;
-  /** Right-aligned hint: the marker the row inserts, or a field's path. */
-  hint: string | undefined;
-};
-
-/** Root rows tagged with a section so the menu can render Notion-style grouped
- *  dividers while keeping one flat, query-filtered, keyboard-navigable order. */
-const SLASH_ROOT_GROUP = {
-  "create-field": "primary",
-  "create-condition": "primary",
-  field: "reuse",
-  "open-fields": "reuse",
-  "open-clauses": "reuse",
-} as const satisfies Record<SlashRootItem["kind"], "primary" | "reuse">;
-
-const SLASH_GROUP_LABEL = {
-  primary: "templates.studio.slashGroupInsert",
-  reuse: "templates.studio.slashGroupReuse",
-} as const satisfies Record<"primary" | "reuse", TranslationKey>;
-
-/** Keyboard-first, command-palette-style menu anchored at the caret, opened by
- *  typing `/` in template prose. The highlighted row tracks the slash plugin's
- *  Up/Down/Left/Right/Enter; clicks use `keepEditorFocus` so the painted caret
- *  stays put while inserting. Three views: a grouped root of creators + submenu
- *  openers, an existing-field reuse list, and a searchable clause library. The
- *  highlighted row's description flies out beside the menu via the shared
- *  preview-pane layout. A pinned footer carries the Escape hint. */
-const SlashMenuPopover = ({
-  slash,
-  rows,
-  highlight,
-  fields,
-  onHighlight,
-  onActivateRoot,
-  onActivateField,
-  onActivateClause,
-  onBack,
-}: {
-  slash: SlashMenu;
-  rows: SlashRows;
-  highlight: number;
-  fields: StudioField[];
-  onHighlight: (index: number) => void;
-  onActivateRoot: (item: SlashRootItem) => void;
-  onActivateField: (path: string) => void;
-  onActivateClause: (clause: SlashClause) => void;
-  onBack: () => void;
-}) => {
-  const t = useTranslations();
-  const empty =
-    rows.view === "clauses"
-      ? t("clauses.noResults")
-      : t("templates.studio.slashEmpty");
-  return (
-    <div
-      className="bg-popover text-popover-foreground absolute z-50 flex flex-col rounded-md border text-sm shadow-lg/5 transition-opacity duration-100 starting:opacity-0"
-      role="listbox"
-      style={{
-        left: slash.left,
-        top: slash.top,
-        transform:
-          slash.placement === "above"
-            ? `translate(0, calc(-100% - ${SLASH_MENU_OFFSET_PX}px))`
-            : `translate(0, ${SLASH_MENU_OFFSET_PX}px)`,
-      }}
-    >
-      <SlashMenuHeader query={slash.query} view={rows.view} onBack={onBack} />
-      <MenuPreviewLayout
-        className="min-h-0"
-        preview={
-          <PreviewPane>
-            <SlashPreview fields={fields} highlight={highlight} rows={rows} />
-          </PreviewPane>
-        }
-      >
-        <div className="max-h-[min(18rem,60vh)] min-h-0 w-56 overflow-y-auto p-1">
-          {rows.items.length === 0 && (
-            <p className="text-muted-foreground px-2 py-1.5 text-xs">{empty}</p>
-          )}
-          <SlashMenuRows
-            fields={fields}
-            highlight={highlight}
-            rows={rows}
-            onActivateClause={onActivateClause}
-            onActivateField={onActivateField}
-            onActivateRoot={onActivateRoot}
-            onHighlight={onHighlight}
-          />
-        </div>
-      </MenuPreviewLayout>
-      <div className="text-muted-foreground border-t px-3 py-1 text-[11px] leading-snug">
-        {t("templates.studio.slashFooter")}
-      </div>
-    </div>
-  );
-};
-
-/** Compact header: the breadcrumb/back affordance for a submenu, or the typed
- *  query / typing hint at the root. */
-const SlashMenuHeader = ({
-  view,
-  query,
-  onBack,
-}: {
-  view: SlashView;
-  query: string;
-  onBack: () => void;
-}) => {
-  const t = useTranslations();
-  if (view === "root") {
-    return (
-      <p className="text-muted-foreground truncate border-b px-3 py-1.5 text-[11px] leading-snug">
-        {query === "" ? t("templates.studio.slashHint") : `/${query}`}
-      </p>
-    );
-  }
-  const label =
-    view === "fields"
-      ? t("templates.studio.existingField")
-      : t("common.clauses");
-  return (
-    <button
-      className="text-muted-foreground hover:text-foreground flex items-center gap-1 border-b px-3 py-1.5 text-start text-[11px] leading-snug"
-      onClick={onBack}
-      onMouseDown={keepEditorFocus}
-      type="button"
-    >
-      <DirectionalIcon className="size-3 shrink-0" icon={ChevronLeftIcon} />
-      <span className="truncate">
-        {label}
-        {query === "" ? "" : ` · ${query}`}
-      </span>
-    </button>
-  );
-};
-
-const SlashMenuRows = ({
-  rows,
-  highlight,
-  fields,
-  onHighlight,
-  onActivateRoot,
-  onActivateField,
-  onActivateClause,
-}: {
-  rows: SlashRows;
-  highlight: number;
-  fields: StudioField[];
-  onHighlight: (index: number) => void;
-  onActivateRoot: (item: SlashRootItem) => void;
-  onActivateField: (path: string) => void;
-  onActivateClause: (clause: SlashClause) => void;
-}) => {
-  if (rows.view === "root") {
-    return (
-      <SlashRootRows
-        highlight={highlight}
-        rows={rows.items}
-        onActivateRoot={onActivateRoot}
-        onHighlight={onHighlight}
-      />
-    );
-  }
-  if (rows.view === "fields") {
-    return (
-      <>
-        {rows.items.map((field, index) => (
-          <SlashMenuRow
-            key={field.path}
-            face={slashFieldFace(field, fields)}
-            selected={index === highlight}
-            onHighlight={() => onHighlight(index)}
-            onSelect={() => onActivateField(field.path)}
-          />
-        ))}
-      </>
-    );
-  }
-  return (
-    <>
-      {rows.items.map((clause, index) => (
-        <SlashMenuRow
-          key={clause.id}
-          face={{ icon: TextQuoteIcon, label: clause.title, hint: undefined }}
-          selected={index === highlight}
-          onHighlight={() => onHighlight(index)}
-          onSelect={() => onActivateClause(clause)}
-        />
-      ))}
-    </>
-  );
-};
-
-/** Root rows, split into Notion-style grouped sections with quiet labels while
- *  keeping the flat highlight index aligned with the key handler's order. */
-const SlashRootRows = ({
-  rows,
-  highlight,
-  onHighlight,
-  onActivateRoot,
-}: {
-  rows: SlashRootItem[];
-  highlight: number;
-  onHighlight: (index: number) => void;
-  onActivateRoot: (item: SlashRootItem) => void;
-}) => {
-  const t = useTranslations();
-  return (
-    <>
-      {rows.map((item, index) => {
-        const prev = rows.at(index - 1);
-        const group = SLASH_ROOT_GROUP[item.kind];
-        const showLabel =
-          index === 0 ||
-          prev === undefined ||
-          SLASH_ROOT_GROUP[prev.kind] !== group;
-        // Field rows carry a dynamic label (the field name), so resolve their
-        // face inline; the fixed entry rows resolve a translation key.
-        let face: SlashRowFace;
-        if (item.kind === "field") {
-          face = {
-            icon: BracesIcon,
-            label: item.label,
-            hint: item.label === item.path ? undefined : item.path,
-          };
-        } else {
-          const rootFace = slashRootFace(item);
-          face = {
-            icon: rootFace.icon,
-            label: t(rootFace.labelKey),
-            hint: rootFace.hint,
-          };
-        }
-        return (
-          <div key={slashRootKey(item)}>
-            {showLabel && (
-              <p className="text-muted-foreground px-2 pt-1.5 pb-0.5 text-[10px] font-medium tracking-wide uppercase">
-                {t(SLASH_GROUP_LABEL[group])}
-              </p>
-            )}
-            <SlashMenuRow
-              chevron={
-                item.kind === "open-fields" || item.kind === "open-clauses"
-              }
-              face={face}
-              selected={index === highlight}
-              onHighlight={() => onHighlight(index)}
-              onSelect={() => onActivateRoot(item)}
-            />
-          </div>
-        );
-      })}
-    </>
-  );
-};
-
-const SlashMenuRow = ({
-  face,
-  selected,
-  chevron = false,
-  onHighlight,
-  onSelect,
-}: {
-  face: SlashRowFace;
-  selected: boolean;
-  chevron?: boolean;
-  onHighlight: () => void;
-  onSelect: () => void;
-}) => {
-  const { icon: Icon, label, hint } = face;
-  const rowRef = useRef<HTMLButtonElement | null>(null);
-  // Keep the keyboard-highlighted row in view when the list overflows.
-  useExternalSyncEffect(() => {
-    if (selected) {
-      rowRef.current?.scrollIntoView({ block: "nearest" });
-    }
-  }, [selected]);
-  return (
-    <button
-      aria-selected={selected}
-      className={cn(
-        "flex w-full items-center gap-2 rounded-sm px-2 py-1 text-start",
-        selected ? "bg-accent text-accent-foreground" : "text-foreground",
-      )}
-      // eslint-disable-next-line react/react-compiler -- containedHandler defers the rowRef.current read into the returned click handler; the ref is not read during render
-      onClick={containedHandler(rowRef, onSelect)}
-      // eslint-disable-next-line react/react-compiler -- containedHandler defers the rowRef.current read into the returned mousedown handler; the ref is not read during render
-      onMouseDown={containedHandler(rowRef, (event: ReactMouseEvent) =>
-        event.preventDefault(),
-      )}
-      onMouseEnter={onHighlight}
-      ref={rowRef}
-      role="option"
-      tabIndex={-1}
-      type="button"
-    >
-      <Icon className="text-muted-foreground size-3.5 shrink-0" />
-      <span className="flex-1 truncate" dir="auto">
-        {label}
-      </span>
-      {hint !== undefined && (
-        <span className="text-muted-foreground shrink-0 truncate font-mono text-[10px]">
-          {hint}
-        </span>
-      )}
-      {chevron && (
-        <DirectionalIcon
-          className="text-muted-foreground size-3.5 shrink-0"
-          icon={ChevronRightIcon}
-        />
-      )}
-    </button>
-  );
-};
-
-/** The flyout description for the highlighted row, shown in the preview pane. */
-const SlashPreview = ({
-  rows,
-  highlight,
-  fields,
-}: {
-  rows: SlashRows;
-  highlight: number;
-  fields: StudioField[];
-}) => {
-  const t = useTranslations();
-  if (rows.view === "clauses") {
-    const clause = rows.items.at(highlight);
-    if (clause === undefined) {
-      return <SlashPreviewEmpty />;
-    }
-    return <SlashClausePreview clause={clause} />;
-  }
-  if (rows.view === "fields") {
-    const field = rows.items.at(highlight);
-    if (field === undefined) {
-      return <SlashPreviewEmpty />;
-    }
-    return (
-      <SlashTextPreview
-        marker={`{{${field.path}}}`}
-        title={slashFieldFace(field, fields).label}
-        body={t("templates.studio.conceptField")}
-      />
-    );
-  }
-  const item = rows.items.at(highlight);
-  if (item === undefined) {
-    return <SlashPreviewEmpty />;
-  }
-  return <SlashRootPreview item={item} />;
-};
-
-const SlashPreviewEmpty = () => {
-  const t = useTranslations();
-  return (
-    <div className="text-muted-foreground flex h-full items-center justify-center text-center text-xs text-balance">
-      {t("templates.studio.previewHint")}
-    </div>
-  );
-};
-
-const SlashRootPreview = ({ item }: { item: SlashRootItem }) => {
-  const t = useTranslations();
-  if (item.kind === "create-field") {
-    return (
-      <SlashTextPreview
-        marker={`{{${item.path}}}`}
-        title={t("templates.studio.makeField")}
-        body={t("templates.studio.conceptField")}
-      />
-    );
-  }
-  if (item.kind === "create-condition") {
-    return (
-      <SlashTextPreview
-        marker="{{#if …}} … {{/if}}"
-        title={t("templates.studio.showOnlyIf")}
-        body={t("templates.studio.conceptCondition")}
-      />
-    );
-  }
-  if (item.kind === "open-fields") {
-    return (
-      <SlashTextPreview
-        marker="{{ … }}"
-        title={t("templates.studio.existingField")}
-        body={t("templates.studio.conceptField")}
-      />
-    );
-  }
-  return (
-    <SlashTextPreview
-      marker="{{@clause: … }}"
-      title={t("common.clauses")}
-      body={t("templates.studio.conceptClause")}
-    />
-  );
-};
-
-const SlashTextPreview = ({
-  marker,
-  title,
-  body,
-}: {
-  marker: string;
-  title: string;
-  body: string;
-}) => (
-  <div className="flex h-full flex-col gap-1.5 overflow-hidden text-xs">
-    <code className="bg-primary/10 text-primary w-fit rounded px-1 py-0.5 text-[10px]">
-      {marker}
-    </code>
-    <p className="text-foreground font-medium">{title}</p>
-    <p className="text-muted-foreground leading-snug">{body}</p>
-  </div>
-);
-
-const SlashClausePreview = ({ clause }: { clause: SlashClause }) => {
-  const t = useTranslations();
-  const description = clause.description?.trim();
-  return (
-    <div className="flex h-full flex-col gap-1.5 overflow-hidden text-xs">
-      <p className="text-foreground font-medium" dir="auto">
-        {clause.title}
-      </p>
-      <p className="text-muted-foreground">
-        {t("common.versionLabel", { version: String(clause.currentVersion) })}
-      </p>
-      <p className="text-muted-foreground leading-snug" dir="auto">
-        {description && description.length > 0
-          ? description
-          : t("templates.studio.conceptClause")}
-      </p>
-    </div>
-  );
-};
-
-/** Stable React key for a root row. */
-const slashRootKey = (item: SlashRootItem): string => {
-  if (item.kind === "create-field") {
-    return `create-field:${item.path}`;
-  }
-  if (item.kind === "field") {
-    return `field:${item.path}`;
-  }
-  return item.kind;
-};
-
-/** The four root-row labels, none of which take ICU arguments, so the caller's
- *  `t(labelKey)` needs no values arg (typing this as the broad `TranslationKey`
- *  would force one). */
-type SlashRootLabelKey =
-  | "templates.studio.makeField"
-  | "templates.studio.showOnlyIf"
-  | "templates.studio.existingField"
-  | "common.clauses";
-
-type SlashRootFace = {
-  icon: LucideIcon;
-  labelKey: SlashRootLabelKey;
-  hint: string | undefined;
-};
-
-/** Icon + label-key + marker hint for a root row, mirroring the gesture-menu
- *  vocabulary. Returns the key (not the resolved string) so the caller's `t`
- *  does the lookup; passing `t` itself around explodes the message-key union. */
-const slashRootFace = (item: SlashRootItem): SlashRootFace => {
-  if (item.kind === "create-field") {
-    return {
-      icon: BracesIcon,
-      labelKey: "templates.studio.makeField",
-      hint: "{{ }}",
-    };
-  }
-  if (item.kind === "create-condition") {
-    return {
-      icon: SplitIcon,
-      labelKey: "templates.studio.showOnlyIf",
-      hint: "{{#if}}",
-    };
-  }
-  if (item.kind === "open-fields") {
-    return {
-      icon: BracesIcon,
-      labelKey: "templates.studio.existingField",
-      hint: undefined,
-    };
-  }
-  return {
-    icon: TextQuoteIcon,
-    labelKey: "common.clauses",
-    hint: undefined,
-  };
-};
-
-const slashFieldFace = (
-  field: StudioField,
-  fields: StudioField[],
-): SlashRowFace => {
-  const match = fields.find((f) => f.path === field.path);
-  const icon = match
-    ? VALUE_TYPE_META[inputTypeValueKind(match.inputType)].icon
-    : BracesIcon;
-  const label = field.label === "" ? field.path : field.label;
-  return {
-    icon,
-    label,
-    hint: label === field.path ? undefined : field.path,
-  };
-};
-
-/** What each gesture-menu insert produces, shown in the preview pane while a
- *  row is highlighted. Mock depictions of inserted document content, not
- *  interface text; deliberately untranslated to keep them free of per-language
- *  i18n debt. */
-type GestureInsertKind = "field" | "if" | "each" | "clause";
-
-// Mock document content, not interface text; deliberately untranslated (see
-// the preview-pane pattern) so the previews carry no per-language i18n debt.
-// The concept caption below each IS translated (templates.studio.concept*).
-const GESTURE_PREVIEW_SAMPLE = {
-  fieldBefore: "Due within ",
-  fieldMarker: "{{term}}",
-  fieldValue: "30",
-  fieldAfter: " days.",
-  ifMarker: "{{#if has_guarantor}}",
-  ifBody: "The Guarantor shall be jointly liable.",
-  eachMarker: "{{#each parties}}",
-  eachItem: "– {{name}}, {{role}}",
-  eachFilled: ["– Jane Roe, Buyer", "– John Doe, Seller"],
-  clauseMarker: "{{@clause:liability}}",
-  clauseText: "Neither party is liable for indirect or consequential loss.",
-} as const;
-
-const CONCEPT_KEY = {
-  field: "templates.studio.conceptField",
-  if: "templates.studio.conceptCondition",
-  each: "templates.studio.conceptLoop",
-  clause: "templates.studio.conceptClause",
-} as const satisfies Record<GestureInsertKind, TranslationKey>;
-
-/** One-shot reveal: fills in after a short beat and stays (no loop). */
-const useFillReveal = (): boolean => {
-  const [filled, setFilled] = useState(false);
-  useMountEffect(() => {
-    const id = setTimeout(() => setFilled(true), 600);
-    return () => clearTimeout(id);
-  });
-  return filled;
-};
-
-/** Cross-fades the whole template form into the whole filled result, stacked in
- *  one grid cell sized to both — so the surrounding layout never reflows and a
- *  shorter filled line just leaves harmless trailing space (no mid-text gap).
- *  Opacity only (GPU-friendly); one-shot, settles on the filled layer. */
-const FillReveal = ({
-  on,
-  marker,
-  filled,
-  className,
-}: {
-  on: boolean;
-  marker: ReactNode;
-  filled: ReactNode;
-  className?: string;
-}) => (
-  <span className={cn("grid [&>*]:col-start-1 [&>*]:row-start-1", className)}>
-    <span
-      aria-hidden={on}
-      className="transition-opacity duration-500"
-      style={{ opacity: on ? 0 : 1 }}
-    >
-      {marker}
-    </span>
-    <span
-      aria-hidden={!on}
-      className="transition-opacity duration-500"
-      style={{ opacity: on ? 1 : 0 }}
-    >
-      {filled}
-    </span>
-  </span>
-);
-
-const GestureInsertPreview = ({ kind }: { kind: GestureInsertKind }) => {
-  const t = useTranslations();
-  const filled = useFillReveal();
-
-  return (
-    <div className="flex h-full flex-col">
-      <div className="flex flex-1 flex-col justify-center overflow-hidden text-sm leading-relaxed">
-        {kind === "field" && (
-          <FillReveal
-            className="whitespace-nowrap"
-            filled={
-              <span>
-                {GESTURE_PREVIEW_SAMPLE.fieldBefore}
-                <span className="text-foreground font-semibold">
-                  {GESTURE_PREVIEW_SAMPLE.fieldValue}
-                </span>
-                {GESTURE_PREVIEW_SAMPLE.fieldAfter}
-              </span>
-            }
-            marker={
-              <span>
-                {GESTURE_PREVIEW_SAMPLE.fieldBefore}
-                <code className="bg-primary/10 text-primary rounded px-1 py-0.5 text-xs">
-                  {GESTURE_PREVIEW_SAMPLE.fieldMarker}
-                </code>
-                {GESTURE_PREVIEW_SAMPLE.fieldAfter}
-              </span>
-            }
-            on={filled}
-          />
-        )}
-
-        {kind === "if" && (
-          <div className="border-foreground-disabled bg-accent/50 rounded-sm border-s-[3px] py-1.5 ps-3 pe-2">
-            <FillReveal
-              className="block"
-              filled={
-                <span className="text-success inline-flex items-center gap-1 text-xs">
-                  <CheckIcon className="size-3" />
-                </span>
-              }
-              marker={
-                <code className="text-muted-foreground block text-xs">
-                  {GESTURE_PREVIEW_SAMPLE.ifMarker}
-                </code>
-              }
-              on={filled}
-            />
-            <span className="mt-1 block">{GESTURE_PREVIEW_SAMPLE.ifBody}</span>
-          </div>
-        )}
-
-        {kind === "each" && (
-          <div className="border-success/40 bg-success/10 rounded-sm border-s-[3px] py-1.5 ps-3 pe-2">
-            <FillReveal
-              className="block"
-              filled={
-                <span className="block">
-                  {GESTURE_PREVIEW_SAMPLE.eachFilled.map((row) => (
-                    <span className="block" key={row}>
-                      {row}
-                    </span>
-                  ))}
-                </span>
-              }
-              marker={
-                <span className="block">
-                  <code className="text-muted-foreground text-xs">
-                    {GESTURE_PREVIEW_SAMPLE.eachMarker}
-                  </code>
-                  <span className="mt-1 block">
-                    {GESTURE_PREVIEW_SAMPLE.eachItem}
-                  </span>
-                </span>
-              }
-              on={filled}
-            />
-          </div>
-        )}
-
-        {kind === "clause" && (
-          <FillReveal
-            className="block"
-            filled={
-              <span className="block leading-relaxed">
-                {GESTURE_PREVIEW_SAMPLE.clauseText}
-              </span>
-            }
-            marker={
-              <span className="bg-card flex items-center gap-1.5 rounded-sm border border-dashed p-2">
-                <TextQuoteIcon className="text-muted-foreground size-3.5 shrink-0" />
-                <code className="text-muted-foreground text-xs">
-                  {GESTURE_PREVIEW_SAMPLE.clauseMarker}
-                </code>
-              </span>
-            }
-            on={filled}
-          />
-        )}
-      </div>
-      <p className="text-muted-foreground mt-1.5 shrink-0 text-[11px] leading-snug">
-        {t(CONCEPT_KEY[kind])}
-      </p>
-    </div>
-  );
-};
-
-/** One reusable item in a {@link GestureSplitRow}'s expand list. */
-type GestureExistingOption = {
-  key: string;
-  icon: LucideIcon;
-  label: string;
-  /** Secondary muted text (e.g. a field's path), shown right-aligned. */
-  sublabel?: string | undefined;
-  onSelect: () => void;
-};
-
-/** A gesture-menu row whose primary click creates something new (the sensible
- *  default); a trailing chevron expands an inline list of existing items to
- *  reuse instead, so reuse is exactly one extra click. Used for both fields
- *  ("Make field" + existing fields) and conditions ("Show only if" + existing
- *  conditions), keeping the two symmetric. */
-const GestureSplitRow = ({
-  icon: Icon,
-  label,
-  reuseLabel,
-  onDefault,
-  onHighlight,
-  existing,
-}: {
-  icon: LucideIcon;
-  label: string;
-  reuseLabel: string;
-  onDefault: () => void;
-  /** Drive the preview pane while the default (create-new) button is hovered
-   *  or focused. */
-  onHighlight: () => void;
-  existing: GestureExistingOption[];
-}) => {
-  const [open, setOpen] = useState(false);
-  return (
-    <>
-      <div className="flex items-center gap-0.5">
-        <Button
-          className="flex-1 justify-start gap-2 font-normal"
-          onClick={onDefault}
-          onFocus={onHighlight}
-          onMouseDown={keepEditorFocus}
-          onMouseEnter={onHighlight}
-          size="sm"
-          variant="ghost"
-        >
-          <Icon className="text-muted-foreground size-3.5 shrink-0" />
-          {label}
-        </Button>
-        {existing.length > 0 && (
-          <Button
-            aria-expanded={open}
-            aria-label={reuseLabel}
-            className="shrink-0"
-            onClick={() => setOpen((isOpen) => !isOpen)}
-            onMouseDown={keepEditorFocus}
-            size="icon-sm"
-            title={reuseLabel}
-            variant="ghost"
-          >
-            {open ? (
-              <ChevronDownIcon className="text-muted-foreground size-3.5 shrink-0" />
-            ) : (
-              <DirectionalIcon
-                className="text-muted-foreground size-3.5 shrink-0"
-                icon={ChevronRightIcon}
-              />
-            )}
-          </Button>
-        )}
-      </div>
-      {open && (
-        <div className="flex max-h-44 flex-col overflow-y-auto">
-          {existing.map((option) => {
-            const OptionIcon = option.icon;
-            return (
-              <Button
-                className="justify-start gap-2 ps-7 font-normal"
-                key={option.key}
-                onClick={option.onSelect}
-                onMouseDown={keepEditorFocus}
-                size="sm"
-                title={option.sublabel ?? option.label}
-                variant="ghost"
-              >
-                <OptionIcon className="text-muted-foreground size-3.5 shrink-0" />
-                <span className="min-w-0 truncate">{option.label}</span>
-                {option.sublabel === undefined ? null : (
-                  <code className="text-muted-foreground ms-auto min-w-0 truncate text-[10px]">
-                    {option.sublabel}
-                  </code>
-                )}
-              </Button>
-            );
-          })}
-        </div>
-      )}
-    </>
-  );
-};
-
-/** The AI slot under the separator: a shimmer while the proposal is in
- *  flight, then the proposal as a quietly accented row — proposed type icon
- *  left, wand right. A proposal that points at an existing field reads as
- *  "Use existing …" and reuses it instead of creating a duplicate. */
-const GestureAiRow = ({
-  enrichment,
-  fields,
-  onAcceptAi,
-}: {
-  enrichment: GestureEnrichment;
-  fields: StudioField[];
-  onAcceptAi: () => void;
-}) => {
-  const t = useTranslations();
-  if (enrichment.status !== "ready") {
-    return (
-      <div className="flex h-8 items-center px-2.5">
-        <span
-          aria-hidden="true"
-          className="bg-muted h-1.5 w-16 animate-pulse rounded-full"
-        />
-      </div>
-    );
-  }
-  const existing =
-    enrichment.fieldPath === undefined
-      ? undefined
-      : fields.find((f) => f.path === enrichment.fieldPath);
-  let label: string;
-  let TypeIcon: LucideIcon;
-  if (existing !== undefined) {
-    label = t("templates.studio.useExisting", {
-      name: existing.label === "" ? existing.path : existing.label,
-    });
-    TypeIcon = VALUE_TYPE_META[inputTypeValueKind(existing.inputType)].icon;
-  } else {
-    label =
-      enrichment.label !== undefined && enrichment.label !== ""
-        ? enrichment.label
-        : t("templates.studio.makeField");
-    TypeIcon =
-      VALUE_TYPE_META[inputTypeValueKind(enrichment.inputType ?? "text")].icon;
-  }
-  return (
-    <Button
-      className="bg-primary/5 text-primary hover:bg-primary/10 hover:text-primary justify-start gap-2 font-normal"
-      onClick={onAcceptAi}
-      onMouseDown={keepEditorFocus}
-      size="sm"
-      title={label}
-      variant="ghost"
-    >
-      <TypeIcon className="size-3.5 shrink-0" />
-      <span className="min-w-0 truncate">{label}</span>
-      <WandSparklesIcon className="ms-auto size-3.5 shrink-0" />
-    </Button>
-  );
-};
-
 /** A paragraph node carrying marker text ("" for the empty body line). */
 const markerParagraph = (
   state: EditorState,
